Ensure that RewardEvent.todict returns only lists (cpu), also set timings correctly for rouge and relevance models

+++ prompting/rewards/rouge_reward.py
@@ -47,7 +47,7 @@ class RougeRewardModel(BaseRewardModel):
 
         output = BatchRewardOutput(
             rewards=torch.FloatTensor(rewards),
-            timings=timings,
+            timings=torch.FloatTensor(timings),
             extra_info={
                 "ngram": self.ngram,
                 "metric": self.metric,
